adding ItemsViewModelTest
